Fix a bug in View.js for re-renders of View's operating in isTemplateMode
A View can be re-rendered. IF it's operating in isTemplateMode, the
re-render does not replace the existing node.

This is not good and is a blocker for fixing the HTML validation
issue inside the DownloadIcon which needs to be re-rendered

A test is modified to include the edge case
which bumps up code coverage on this file

Bug: T205359
Change-Id: I549e15f8a3352761e4960abdaaef6cd4bfb133e6