Update patch set 2
Patch Set 2:

> looks like the branch on line 49 inside the show method is not
 > being tested. Can we add a test showing that two calls to show keep
 > the panel visible

Line 37 of newest patch has new test for calling show() when the panel is already shown. I think what I didn't like about this (and why I left it out of my previous patch) is that, to make the test have value, it uses a setTimeout and panel.minHideDelay inside the test (if you don't use a setTimeout then the assertions may fire prematurely) which felt pretty coupled to Panel's implementation. If you know of a better way to test this, please let me know!

Patch-set: 2