Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

Mandatory disclaimer: I'm not overly familiar with MF, so I'm only commenting on the things I know (or think I know!).

The -1 is due to the usage of ->text() on a Message object if -- and when, in this case, as I understand it -- it isn't strictly needed. Based on some codesearch results, skinname-<skin name> messages are always simple strings (e.g. no {{GRAMMAR:}} or other magic there) so I can't see why we'd want to allow raw HTML parsing, even in theory. This can be and I hope it is safe as IIRC the Special:Preferences form uses HTMLForm, which should take care of proper escaping and whatnot, but as a reviewer I strongly prefer code to be demonstratably safe, and in this case swapping text to escaped does that and reduces my anxiety. :)

Patch-set: 3
Reviewer: Jack Phoenix <243@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1