Separate layout rules of spinner styles to content area
The spinner is always used in a content area. In the only place it
isn't - the NotificationBadge - and here the rule has no impact due
to the fixed width of the icon.

I would like to use the spinner in the DownloadIcon. In this context
the margin can cause problems in the rendering (repaints) so I'd like
to scope these to .content (see Ib87390d17bef6f50842f52cd84c9ce2b162aaff0)
This is one of the blockers for doing that.

Two rules are needed given the EditorOverlay and other overlays due to
not always mark up the overlay-content element with content class

Change-Id: I4d703eef68d51bbe0b03579c5cca0845e17b8c9d