mobile.startup/Overlay: Compatibility with skins other than Minerva
* Append itself to 'body' instead of '#mw-mf-viewport'
  (Jon says that the reasons to use that are no longer valid)
* Add necessary styles that only existed in Minerva

Running `mw.loader.load( 'skins.minerva.editor' )` will now load the
mobile editor on any skin in mobile mode, and it will actually work and
be able to edit pages!

Many things still look obviously broken when using it (a bunch of styles
for the toolbar and buttons seem missing), but it kind of works, and
the rest should be easy to track down (if perhaps tedious).

Bug: T198765
Change-Id: I1c1572e572826dfd9f94a38af9e05ee70c015d26