Update patch set 3
Patch Set 3:

(1 comment)

As mundane as it might seem, let's add a new test for this as well.
Before this change OverlayManager code coverage was 98-100% and now it's down to 93.

One of the team's goals right now is to improve and enforce coverage in MobileFrontend.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>