Remove unreachable code and add more test cases + docs
Edge cases also doesn't reach this code such as NULLs, empty
string, and a given URL without a scheme. No matter the values of
$this->wmfWikiHosts and $this->wmfMultiDomainWikiHosts, a count on
the sub-domain called by extractSubdomain() method will always default
to 1 after a call to explode().

Example; print_r( count( explode( '.', 'MediaWiki' ) ) ); will return 1

That is the expected behavior as if no delimiter is found, it will
always return an array with 1 element (the string passed as argument).

Lines of code removed are;
* return $baseDomain; (removed L#100 - L#102)

Code coverage at 96.00%

Bug: T209218
Change-Id: I7917bdac3ca0149ba43bb887486d2aae23103c38