Fix $expected and $actual values in MobileCollectionTest
So $expected and $actual was used in the reverse sense. As MobileCollection
is the main SUT with it's methods ::count(), ::add(), ::getIterator() to be
tested, their return values should be the $actual not the $expected.

This is a follow-up patch for: I026110cf13f23fb4a492e773b926ed0737dea0d4.

Bug: T208987
Change-Id: Ie93a235d09ce0d00546cce1d58cf9763ef4da59c