Update patch set 11
Patch Set 11:

@jdlrobson, there is an alternative approach to solving this issue without removing the unreachable code if we really don't want to remove it.

We can use the @codeCoverageIgnoreStart and @codeCoverageIgonoreEnd to make PHPUnit ignore this lines of code when generating the coverage report. So we'll still have the code, more test inputs (of course), and coverage will increase (likely still hit 100%) but with the annotations around the parts/lines of code (unreachable) that will be ignored.

So should we go for this approach or we maintain this patch? Thanks :)

Patch-set: 11
Reviewer: D3r1ck01 <2842@e9e9afe9-4712-486d-8885-f54b72dd1951>