Add test coverage for rlModuleLoader.js
rlModuleLoader.js now has 100% test coverage. Before this commit it had
0% (no test file);

Notable changes:

* A new property (newLoadingOverlay) was added to rlModuleLoader.js
exports to facilitate stubbing LoadingOverlay in the tests.

* It was decided to use spies for "LoadingOverlay's" show/hide methods
because testing the output ($el property) of the real LoadingOverlay was
too awkward/tedious and did not feel within the responsibility of
rlModuleLoader. The real desired behavior of rlModuleLoader is calling
.show and .hide at the appropriate times on a loadingOverlay object
thus the tests test for that.

Bug: T209185
Change-Id: I818cd844eeef51cd2bf22ed3a995d5c9d85dc572