Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Thanks for removing these closures. I'm all for that!

On the other hand this seems to break the linting we setup to contain jQuery to only be usable in certain classes (as part of a painful push we made to limit our jQuery usage in favor of es5/es6).

The team invested a lot of time limiting jQuery usage to avoid access to global jQuery. This undoes much of that work.

Specifically, if I replace
```
				this.$( '.page-list' ).removeClass( 'hidden' );
``` 
in resources/mobile.nearby/Nearby.js

with
```
$( '.page-list' ).removeClass( 'hidden' );
```
and run
```
npm run lint:js
```

It's supposed to throw an eslint error to tell me not to use jQuery.

After your change this no longer happens.

Please retain the previous behaviour.

I've added a few comments explaining how we can achieve your goal but also limit jQuery to where it's allowed to be accessed as a global.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1