Update patch set 4
Patch Set 4:

(1 comment)

> @Anomie I'm happy to follow up these changes with such refactors if Pjht is reluctant to do so

At the very least, the check for keys beginning with "_" should be replaced with use of ApiResult::isMetadataKey() before this is merged. As written it's bypassing abstraction to depend on internal details (details which I'd like to change someday, BTW).

The rest is just not as good as it could be and would be ok as a followup.

Patch-set: 4