Update patch set 6
Patch Set 6:

D3r1ck01: please check my patch. This is a sample, most probably we will remove the `integrations/*` code in the future (please  see the task I created, there is a todo in code). I'm not sure if that extra `PageImages::getPageImage()` call is required, looks like not but I want to take a closer look.

I made this patch as an example how we can implement this logic in a nice way. I used the Dependency Injection pattern (see ServiceWirings.php) to provide us correct version of the PageImageProvider. There are two classes that provide the page image, first one that uses PageImages extension, and second one (Null Object pattern) that returns false as there is no way to retrieve the page image. MobilePage becaame stupid class that always takes $file  and doesn't try to fetch it magically. The SpecialMobileEditWatchlist is responsible to pass the correct $file all the time.

This patchset introduces couple files and two patterns - NullObject and DependencyInjection. It may seem complicated but overall it makes classes dumb. Each class should be reponsible only for one thing. For example SpecialMobileEditWatchlist asks for the `MobileFrontend.PageImageProvider` service, it doesn't care whats that, it doesn't know about the PageImages extension, it just retrieves the service and calls `fetchImage()`.

Patch-set: 6
Reviewer: Pmiazga <3945@e9e9afe9-4712-486d-8885-f54b72dd1951>