Update patch set 19
Patch Set 19:

> This solution looks pretty neat. Nice use of the services/mocking and thank you for the descriptive commit msg.

Kudos to @Pmiazga for the helping hand ;)

> I'm seeing 98.18% coverage.. which bit of the class is not covered out of interest?

Thanks for your reviews @jdlrobson and yes, I'm figuring out which bit of code is not yet tested, I'll need to generate the code coverage locally to see exactly which bit is not and fix that. In that light, another patch set is coming up soon which will complete the coverage. WIP :)

Patch-set: 19
Reviewer: D3r1ck01 <2842@e9e9afe9-4712-486d-8885-f54b72dd1951>