Remove TalkOverlayBase and a level of inheritance
We've identified our use of inheritance in MobileFrontend as
an anti-pattern. Now thanks to the className option in View's
there is no need to do this for TalkOverlay.

A whole class gone!

While this duplicates setup of the pageGateway and editorApi
and passing the className to the constructor this removes a
level of indirection which makes the code more readable and
predictable.

Changes:
* Autosign is not a member of TalkOverlayBase class. It's now
its own function
The autosign method doesn't require anything inside the class
so can be exported and imported without the class.
* This means that a TalkOverlay can simply extend TalkOverlayBase

Change-Id: I224f53033121b998ac7d21707c90a969d7aca68f