Update patch set 2
Patch Set 2: Code-Review-1

> In eslint-config-wikimedia 0.9.0 we introduce the wikimedia/jquery ruleset. This ruleset defines the '$' global, which can't be overridden back to disallowed, so going forward you would also not be able to take advantage of these jquery rules.

For jQuery projects, the new rules must be as indispensable as other ESLint rulesets have proven to be but given MobileFrontend's push to minimize dependencies, this seems like a step backwards. Maybe we can use this composition [0] in the short term and switch to [1] with per file enables when [2] is fixed?

[0] https://github.com/wikimedia/eslint-config-wikimedia#a-typical-front-end-project
[1] https://github.com/wikimedia/eslint-config-wikimedia#jquery
[2] https://github.com/eslint/eslint/issues/9940

Patch-set: 2
Reviewer: Niedzielski <2426@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1