Update patch set 2
Patch Set 2:

> Ed wrote:
> can't be overridden back to disallowed

This is indeed true, but to be clear, this isn't because we explicitly wanted to disallow that. It's because ESLint doesn't support an ability to turn them off locally (unlike all other aspects of ESLint, which can). See https://github.com/eslint/eslint/issues/9940 for more details about this issue.

Our choices were:

1. Add "$ to wikimedia/client for convenience (so nobody has to enable it), with the downside of it also being allowed in projects that don't or can't use jQuery, and no way to disallow use in projects like MF.

2. Don't add "$" anywhere in the presets, so that when using wikimedia/jquery, you also have to manually enable the "$" global.

3. Add "$" only to the wikimedia/jquery preset. So that it's there for convenience when you're using this preset.

Choices 1 and 2 didn't make sense to us, but choice 3 seems reasonable and from what I understand, would work well for MF's coding principles as well.

If I understand correctly, MF's current practice is to treat jQuery similar to a module that is conditionally loaded (like Moment.js, Mustache, or some other upstream lib). In that, you don't want to allow its global to be used just anywhere, instead only in places where someone intentionally decided to whitelist it. Be it for a specific line of code, a file, or a directory, also with the knowledge that the code in question has the dependency declared.

As such, MF would probably be better off not enabling the wikimedia/jquery preset project-wide. Instead, you'd require it to be explicitly whitelisted via /* global $ */. This would effectively be an extreme version of the no-* rules for jquery, that disallows any and all use of jQuery, unless explicitly opt-ing in.

If you're wanting additional feedback and guidance from ESLint about anti-patterns within the scenarios where it was decided to whitelist jQuery, one could also enable the handful of rules from https://github.com/wikimedia/eslint-config-wikimedia/blob/d41a42fec4cac89891b82786964fe8ade197a5f6/jquery.json#L7-L13 in eslintrc.json, which would achieve the same effect.

Patch-set: 2
CC: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>