Update patch set 3
Patch Set 3: Code-Review+2

> There's an explicit Task about what's happening and why.
We merge hundreds of commits over various codebases in a day. I read the git logs for all of them as much as I can.
When I read a git log, I expect that to be enough.. however often our commit messages are minimal which is a bad
thing I'd like us to get better at. We shouldn't be expected to jump into phabricator (sometimes on tasks with a variety of activity) for explanations to every codebase like on this one. Communication is important in shared codebases.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2