Hygiene: add tests for CtaDrawer
CtaDrawer implements only render-specific deviations from its
superclass, Drawer. This patch adds a test for that unit of deviation.

Unfortunately, the HTML rendered isn't pretty but it is tested. A RegExp
is used to reduce some brittleness in a practical way but it isn't
foolproof.

CtaDrawer functionality was refactored as needed for testability. This
includes extracting functions from preRender() and exporting them. The
exports should not change how the bundle is optimized as they are not
shared outside of the entry.

Some additional opportunistic improvements were made to improve the
JSDocs.

Bug: T208920
Change-Id: Ifaa3bd99c81b0eecf33f731ad5347e0872340bb1