Update patch set 3
Patch Set 3: Code-Review-2

> The problem with this is that you don't get the jQuery plugin rules from the config, so on pages where "/* global $ */" is enabled, users could then use "$.each" etc

Im not understanding why this is a problem. The files where jquery is allowed is being kept to a minimum and we are actively reducing the numbers where it is used and soon it will be limited to 1 or 2 places.

Adding a global $ in new code will be an instant red flag since we are reducing jquery usage and any changes to introduce things like $.each will be picked up by humans rather than eslint in review.

In short discouraging jquery alogether is our goal not discouraging certain functions inside jquery

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2