Remove usage of global event emitter from MF
WARNING: This patch depends on Minerva change
I005d2fcdbf91c2f1ac98178dfa388aa8174e7530 . They should be merged
together.

This patch removes usages of M.on/M.off/M.emit (functionality derived
from moduleLoader.js) from a variety of components in MF that previously
used moduleLoader to listen for or emit these events:

* before-section-toggled
* section-toggled
* scroll
* scroll:throttled,
* resize
* resize:throttled events
* category-added

This patch instead introduces an event bus singleton
(eventBusSingleton.js) that is imported in init scripts at the edges
(not components themselves) and injected into the components that need
it. While the usage of a singleton is not ideal, I think it is warranted
right now as files in both Minerva and MF listen/emit the same set of
events and the refactoring work involved in making a single entry point
from which we then pass down dependencies was deemed out of scope for
this task.

In order to introduce this dependency injection, the jsdoc for the
involved contructors were modifed to account for the eventBusSingleton
being injected and required. There were two very notable changes to the
following constructors:

* ScrollEndEventEmitter's constructor was modified to take the event bus
as a first paramenter.
* Toggler's construtor was modified to take one
object of arguments (previously it took several params)

Additionally, Toggle's ability to act as an event bus was removed. It
used to emit the 'toggled' event on itself, but I could not find
anything listening to this event.

Bug: T156186
Change-Id: Id990b0e1a53221d5c1cb3e3012aed0e27d801fc2