Update patch set 5
Patch Set 5:

I haven't tested this change but it seems reasonable given the discussion 👍 

Would you mind terribly if we waited to merge this patch? @jdlrobson's on holiday and I know he cared quite deeply about this work and would probably appreciate the opportunity to weigh in on it once more before it's merged as it touches a lot of code he authored and is on the hook for maintaining. I've voted -1 to indicate my hope that we wait.

Patch-set: 5
Reviewer: Niedzielski <2426@e9e9afe9-4712-486d-8885-f54b72dd1951>