Update patch set 1
Patch Set 1: -Code-Review

Posting Stephen's question any my reply from I9739ee463c41799f7b220b7b4649a7926dcdee6a 
here for others to see:

> What were your thoughts on picking a singleton pattern? I have no
 > qualms with making a single instance of something but obviously we
 > wish to avoid globals.
 > 
 > I was thinking this could be a parameter passed in which would let
 > us move the singleton to mobile.startup.js and eliminate its usage
 > as a global. However, I guess we'd have to do some big refactoring
 > as soon as we move the code to source and drop the IIFE which would
 > be a task in itself.
 > 
 > I was just hoping to hear your insights on the approach taken. The
 > commit message talks about injecting the dependency which I
 > normally associate with passing from the point of instantiation to
 > every consumer. This seems like a hybrid approach where we pass the
 > dependency when it's practical to do so and use the global
 > otherwise.

I don't particularly like the singleton pattern, but I felt it was appropriate usage here (and other places) as this file (and other "initialization" scripts (files that initialize objects) haven't been ported to webpack and the refactor work required to do that felt out of scope for this task. Additionally, we have initialization scripts in both MF (such as this one) and in Minerva (e.g. skins.minerva.scripts/init.js) that need to share an event bus because they currently listen to or emit the same events. 

I see these patches as a good first step for our event emitter situation, but not the final step that should be taken. I tried to push out the reference to to the singleton as far out as pragmatically possible right now to the initialization scripts (turns out there are many!) and from there on use DI to pass around the event bus. Your comment made me realize that I think a comment is warranted in the eventBusSingleton.js file that it should be judiciously imported (not in component files but in files responsible for initializing the components) which I think sets up well for later refactoring.

Patch-set: 1
Reviewer: Nray <6176@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0