Update patch set 11
Patch Set 11: Code-Review-1

> Patch Set 10: Code-Review-1
> 
> (1 comment)
> 
> Just had one question. I think the approach you've shown here is interesting and I'd be willing to give it a shot. I still have some concerns about the maintainability of our test files doing it this way (I fear we will have to make frequent and cascading changes to them) but I do like the assurance that testing the html string gives. I just wish that our testing harness had something like Jest's snapshot feature where we could easily make snapshots of our component's html output and could also easily regenerate them when needed. The easy to regenerate part I think is key to maintainability of these. Then I think we could have the best of both worlds - assurance that our outputs are what we want and maintainability. That concern is not a blocker for merging this though.

I'm open to alternatives. My intent with this patch is to version a known good HTML render of a component for future testing. If it doesn't work, we delete the test or try something new.

PS11 is just a rebase. I'm working on the inline feedback in a new PS.

Patch-set: 11
Reviewer: Niedzielski <2426@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1