MediaWikiServices & Service Wirings in MwApiContentProvider.php
Created a wrapper method fileGetContents() which is an abstraction of
file_get_contents() in-built function in PHP to use MediaWikiServices and
the HttpRequestFactory class to fetch content via the MwApiContentProvider.
Per above, it opens up a means for DI hence making testing relatively easy.

In addition, use the FormatJson class and it's ::decode method to decode
the content to an array for processing by the getHTML() method in
MwApiContentProvider.

Bug: T212408
Change-Id: I8feee3c6eff91f72d04c1b48604b2dd533240515