Mobile VE should log saveFailure events
This is slightly more complicated than saveSuccess, as we need to log the
reason for the failure.

Incidentally rewrite most of the wikitext editor error handling so that some
of its API response parsing can be reusable. (Also because it looked like
bitrot over the years had taken its toll.)

Bug: T206050
Bug: T202374
Change-Id: I77de9d6b159d66a704ef522e73bd33e9f65893ee