Update patch set 1
Patch Set 1: Code-Review-1

I don't like the hard-coding of the extension functionality. We have similar code in ve.init.mw.ArticleEvents, but at least that is using the events that the extensions are firing, and not repeating the logic for working out if the extension was triggered. Is it possible to re-use code from there, or at least the events (onSaveErrorSpamBlacklist etc.)?

Patch-set: 1
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1