Update patch set 3
Patch Set 3:

(1 comment)

> npm run test:unit also is failing for me when run locally..

The reason the tests are failing is because of a problem with View.test.js and not the tests in toggle.js. This patch just exposes the problem. https://phabricator.wikimedia.org/T210517 

Since the unit tests are non-voting, we could fix this issue in a followup patch, or wait until it's resolved to merge this one.

Patch-set: 3