Bundle mobile.languages.structured with webpack
The mobile.languages.structured module is our first lazy loaded module
to port to webpack. As such, a number of different attempts were tried
to make this work which are summarized at the bottom of this message.

Notable changes:

* mobile.common.js chunk was added to the mobile.startup module in
extension.json because this chunk contains modules shared by
mobile.startup and the language overlay lazy loaded chunk
(mobile.languages.structured.js). Ideally, this chunk would be merged
into the mobile.startup.js and would not exist, but I could not figure
out a way to make webpack do this without also using webpack's lazy
loading feature (read point 2 of "Attempts tried" for an explanation of
why I did not pursue doing that in this patch).

* util.test.js was previously one big test, but I split it up into
smaller tests to improve readability.

* mobile.languages.structured module was added as an explicit dependency
in MobileFrontend.hooks.php because the browser qunit tests were not
loading the module.

Attempts tried:

1) Making a new entry point in webpack.config.js specifically for the
languages overlay and using webpack's SplitChunksPlugin [1] to split
shared modules between mobile.startup and language overlay into a new
common chunk. This was the method chosen by this patch and proved to
have the least amount of friction. The new language overlay chunk will
be lazy loaded by resource loader while the mobile.startup &
mobile.common chunks will be loaded at startup (also by resource
loader). However, additional investigation and work will be needed if
async chunks are added that share modules with other async chunks but
NOT with the mobile.startup chunk. In that case, the modules common
between the async chunks would be excised into the mobile.common chunk
which would increase our critical js size. If this were to occur, a
workaround might be possible using webpack's cacheGroup.test property
[2].

2) Using webpack's lazy loading [3]: I think this is the ideal solution,
but there were a number of hurdles we need to overcome before we can use
this feature. Webpack's dynamic import function makes use of es6
promises so we would need to add a polyfill for browsers that don't
support this. This might be possible using jQuery Deferred's and webpack
[4][5]. The webpack runtime needs to call a public url to the lazy
loaded asset which is configured with the publicPath property [6]. One
can hardcode a path like `/w/extensions/MobileFrontend/resources/dist`
to the dist folder, but this path is configurable through
LocalSettings.php and can change. To get around this, it might be
possible to use webpack's "on the fly" public path [7]. Finally, I had
caching concerns with whatever endpoint we use (e.g. will there be a
lack of caching / will there be excessive caching and how will this play
with RL caching rules). Because of these issues, I forego trying to make
lazy loading work in this patch. More investigation is needed.

3) Using the chunk produced by webpack's lazy loading feature with
resource loader:  As in point 2, if we use webpack's dynamic import
syntax `import()` inside mobile.startup (or any other entry point),
webpack will automatically produce a new chunk on the next build to be
used for lazy loading. In this attempt, however, I tried to use that
chunk with resource loader's lazy loading implementation (and not use
webpack's lazy loading implementation) to avoid concerns discussed in
point 2. However, this chunk is wrapped inside a function with arguments
that can only be supplied by the webpack runtime scope and is only meant
to be executed by the webpack runtime code. Therefore, I could not find
a clean way to make this work.

[1] https://webpack.js.org/plugins/split-chunks-plugin/
[2] https://webpack.js.org/plugins/split-chunks-plugin/#splitchunks-cachegroups-cachegroup-test
[3] https://webpack.js.org/guides/lazy-loading/
[4] https://webpack.js.org/guides/shimming/#loading-polyfills
[5] Stephen
[6] https://webpack.js.org/guides/public-path/
[7] https://webpack.js.org/guides/public-path/#on-the-fly

Bug: T210210
Change-Id: I1f1c73aefd0a7c501773c6261f253807ecd667de