Update patch set 1
Patch Set 1:

> I still think it's better to update them all in one patch.

This patch is for MobileFrontend extension but the parent classes are in the core, here is a PS for the update: Ie8c551a3e8e2d74f52e30a.

So if they're to be merged to keep the consistency, I can make this PS depend on the core's PS right?

Patch-set: 1
Reviewer: D3r1ck01 <2842@e9e9afe9-4712-486d-8885-f54b72dd1951>