Make languages's util.js mirror style of mobile.startup/util.js
This commit changes mobile.languages.structured/util.js so that it
exports an inline object instead of using an intermediate variable
(which is how mobile.startup/uti.js does it). Additionally, the
`@memberof` and `@instance` tags were removed from the
'getDeviceLanguageOrParent' function as it is not a member of the
object.

Bug: T210210
Change-Id: I1659cd66422f36d874d5acdbef8f0cf788ede5a9