Generalise api proxying - Introduce actionParams helper
Make sure all API requests get the default properties - action and
format (can be overriden if necessary) and if running in development
mode the origin parameter is also setup.

This DRYs up a bunch of API requests for readability as well as
making localhost work with production APIs for the read-only version
of all our features.

Bug: T212518
Change-Id: I386494a49b08ce056442639ca7777d3a5c91dfd3