Update: make Webpack build products stable
For consistency with Popups and the rationale posited inline, replace
numerical file references in the bundle with filenames. This keeps
outputs stable so that code only impacts the bundle it lands in instead
of unrelated bundles. This change comes at the expense of increased
bundle sizes.

Also, move noEmitOnErrors up a few lines for consistency with Popups.

Bug: T212527
Change-Id: I5171fdb966b18dfa5a177c0454a33547f5d32958