Drop Icon and Button partial usages outside those classes
Using template partials in the way we do is extremely wasteful
and unnecessry (see T149909). Quite frankly it's embarrasing the
amount of indirection and ways to render Icon and Buttons.

A better way is to make use of postRender statements instead as
it only creates a single DOM element for the associated Icon
(rather than the current approach which creates a DOM element,
only to ignore it and use associated options made via the render
function). Note this is not about getting rid of templatePartials
- there are perfectly valid use cases of this such as in PageList
- this is about not using templatePartials of already existing
components.

A grep for the following results 0 results after this change:
* ">button"
* "Button.prototype.template"
* "Icon.prototype.template"
* ">icon"

The only place these templates will be used is inside Button and
Icon classes

Bug: T149909
Change-Id: I2f0c04e6843f1e4ffc7133e1e5efbd30a27881c1