Streamline McsContentProvider constructor and tests
This is motivated by the changes made in Id7aef17. Most notably, this
makes the new test case much more robust. Now the StatusValue::newFatal
can be changed to newGood and the test will fail, as expected. Before it
was executing an other, unexpected code path and still succeeded.

I also realized the McsContentProvider does not do anything with the
OutputPage object it expected before. Only a Title is needed. With this
more narrow interface a whole bunch of test setup is not needed any more.

I made sure all this is private (all IContentProvider classes are
implementation details of the factory) and does not need to follow any
deprecation policy:
https://codesearch.wmflabs.org/search/?q=McsContentProvider

Bug: T210390
Change-Id: I67ff3fc8e8134e56dfe09d903b802844a9c4812c