Hygiene: extract lazy image loading from Skin
Move lazily loaded images from Skin.js to a new file,
lazyImages/lazyImageLoader, and similarly for the tests. It appears that
Toggler callbacks may need the ability to load images, so isolating this
code is not just a separation of concerns improvement but appears to be
necessary to replace the event bus.

A subdirectory is used as an additional transform object for sections
that will retain loaded image state is planned to be excised from
Skin.js in a future patch.

The code could probably be rewritten to not use jQuery and other
improvements are possible. The tests could also be improved. However,
this patch attempts *minimal* additional changes except to rename
loadImagesList() to loadImages().

Bug: T211724
Related: Ia708cda688e6bdb12074d85d98f7e98fdf7b0ca8
Change-Id: Ic73f78825eaab561e8ed694aa6cc102ccb471f95