Update patch set 5
Patch Set 5:

> I'm struggling to map the goal of the task T209129 to this
 > patchset. MY understanding was the Drawer would no longer be
 > responsible for appending itself but would provide a method that
 > does that. However we seem to be doing the appending in an event
 > now. I don't think events should have side effects.

Ha, ok. My interpretation was that drawers should not append to the DOM *on instantiation*. So calling new Drawer(), should not have the side-effect of appending a new DOM element. The reasoning behind moving this behaviour to the show event is that that's the last point at which it's absolutely. Same goes for the hide event, the DOM gets removed at that point because it's no longer needed.

Patch-set: 5
Reviewer: Jdrewniak <2950@e9e9afe9-4712-486d-8885-f54b72dd1951>