Introduce the AMC opt-in toggle and base AMC classes
The AMC mode is an ongoing experiment, we may or may not keep it.
We are also considering beta features removal. After many tries
I decided to keep the AMC classes separate. With respect to Single
Responsibility Rule we shouldn't mix both beta and AMC modes together
as those are totally separate feature sets. All AMC related features
will stay in the MobileFrontend/AMC namespace. If required, in the
future we can provide a facade - something like
MobileFrontend/UserModeFacade that will provide an easier and unified
access to beta/amc checks.

We cannot make the AMC a single feature (implement the IFeature)
because it's not a single small feature. AMC Manager and UserMode
provide a way to retrieve global on/off state and a way to store user
preference (opted-in and opted-out).
Both Beta mode and AMC mode are just a way to group multiple features
into well defined sets. User can decide which sets he wants to enable
or disable.
FeatureManager is designed to manage all features and it's possible to
handle both beta and amc features. When developing new AMC features,
we are able to register and manage them via FeaturesManager (just
create new config `amc` (we will have 3 configs - base, beta, amc)
plus we will need to provide an AMC feature class that implements
IFeature. With that in mind it will be possible to:
 - have set of AMC features, that can be enabled/disabled via config
 - have a set of Beta features that can be enabled/disabled
 - we will be able to enable subset of Beta features as AMC features
 (one feature can belong to multiple modes)
 - user can opt-in/opt-out to any set of features (beta, amc, both)

Changes:
 * Introduced MobileFrontend/AMC namespace to collect all AMC
   related classes.
 * Introduced Manager, simple class to detect feature availablity.
   Future place for all global AMC related logic
 * Introduced UserMode, an User wrapper to provide easy AMC
   isEnabled() setter and getter, future place for all user related
   logic regarding AMC mode
 * Introduced config variable to disable AMC mode (safety switch)
 * Added new option to Special:MobileOptions that allows enabling
   and disabling the AMC mode for logged in user
 * DRY - moved all OO.infuse related hacks into infuseToggle
   function, as right now they have to applied twice
 * Added AMC/Hooks to register default user option (the default
   value is not stored on user options save)

Bug: T211197
Change-Id: I2dcf79bb8b0ebcc71dafc6981459605ede78e057