Build mobile.special.watchlist.scripts with Webpack
Notable changes:

* The mobile.watchlist module was merged into the
mobile.special.watchlist.scripts module in extension.json as it was the
only thing using it.

*  watchlist was added as a new entrypoint in wepback.config.js AND was
enabled for code splitting  (along with our lazy loaded chunks +
mobile.startup). Enabling code splitting for watchlist benefits new
users to the page that already have the mobile.startup module cached in
the browser as resource loader will only have to request the
watchlist chunk (which doesn't contains any of the mobile.startup
code). If we didn't enable code splitting, the user could be forced to
redownload all of the mobile.startup modules used by the watchlist as
they would be present in the watchlist chunk.

* The downside of enabling code splitting for watchlist is that modules
that are shared by lazy loaded or special page chunks and are NOT in
mobile.startup will be excised into mobile.common.js and increase our
critical js size.

* WatchListGateway's test names were slightly tweaked following our
test conventions [1].

* Multiple stubs for mw.config and mw.util were introduced in
WatchListGateway.test.js to fix discrepencies in browser/headless tests that
caused failures.

[1] https://www.mediawiki.org/wiki/Reading/Web/QUnit_test_guidelines

Bug: T212823
Change-Id: Ib82b240c2582d214cce32fee5832bd28ecd21879