VisualEditorOverlay View code is merged with EditorOverlay
The mobile.editor.ve module is reduced to a single file
ve.init.mw.MobileFrontendArticleTarget.js
The other code is merged into mobile.editor.overlay

This removes an unnecessary module and groups similar code together.
It also means we can remove the mobile.editor.common module in a second
pass.

This change increases the size of mobile.editor.overlay to 6kb (after GZip)
up from 5.3kb.
The majority of this code could be recouped by
with the merging of mobile.editor.overlay with
mobile.editor.api and mobile.editor.common which is now possible
and the refactors that enables.

Additional changes:
* The unused message "mobile-frontend-page-edit-summary" is removed
This was previously the heading of the preview screen but that has not
been the case for some time.

Bug: T207787
Change-Id: Ia6ed40123a2d7ef256961a31b069e08cca6522fa