Hygiene: reduce jQuery in lazyImageLoader
Refactor lazyImageLoader to eliminate a jQuery parameter dependency.
The now unnecessary parameter, _find, will be remove in a future patch.

jQuery is, generally speaking and like most technologies, unwanted when
unneeded:

- jQuery increases the deliverable size. While this change in itself
  won't eliminate jQuery as a dependency it does reduce the overall
  dependency.
- Why add a layer of indirection when the vanilla DOM suffices?
- jQuery is losing popularity and repositories favoring it may also
  appear outdated.
- MobileFrontend has already taken great measures to reduce jQuery
  dependencies to util and View. Even the linting rules forbid casual
  jQuery usage.

In this case specifically, the only reason the effort to minimize jQuery
usage was made was because threading it as a dependency reduced the
legibility of the code.

For consistency, the conversion is completed at file scope in
I85cb520d053e1c6754c8deffb54508db97c50d6b. The work wasn't completed
in this patch as it was too large to consider with other changes.

Bug: T214658
Change-Id: Id8f04ec74eff391a94eb32f48226cbb46bdb1d13