Hygiene: consolidate section-toggled events
Merge before-section-toggled and section-toggled events. The former is
emitted one statement prior to the latter with similar data. The
statement between doesn't appear to impact either subscriber:

  $content
    .toggleClass( 'open-block' )
    .attr( {
      'aria-pressed': !wasExpanded,
      'aria-expanded': !wasExpanded
    } );

Additionally, nobody actually uses the event's section number so drop
that and `expanded` is trimmer than `wasExpanded` at no loss of clarity,
so rename that.

Bug: T211724
Change-Id: Ibb25a97779b966c36821a8c6b7aa7abf3c99faa9